Cleaned up Stepper6600
